This PR introduces an index in the `ATHT_ACCESS_TOKEN` table that is then used when looking for a specific Access Token, specifically during REST API authentication check and token refresh

**Related Issue**
This should really help #1537 but it's not 100% sure that this is the only cause of that issue. We may think to close it after we gather some feedback from the community

**Description of the solution adopted**
Changing the query we do when looking for a specific AccessToken, filtering out already expired and invalidated tokens, will allow the DBMS to use the newly created index to perform the search, cutting down the search time.

**Screenshots**
This from a brief test on my local machine:

Without the index and the optimized query:

```
kapuadb> SELECT * FROM atht_access_token WHERE token_id = '...'
[2020-07-09 11:33:05] 1 row retrieved starting from 1 in 6 s 714 ms (execution: 6 s 676 ms, fetching: 38 ms)
```

With the index and the optimized query:

```
kapuadb> SELECT * FROM atht_access_token WHERE invalidated_on IS NULL AND expires_on > NOW() AND token_id = '...'
[2020-07-09 11:32:29] 1 row retrieved starting from 1 in 60 ms (execution: 13 ms, fetching: 47 ms)
```